Decompose server/src/lib.rs into focused modules (issue #63)
Split the 3890-line lib.rs into state.rs, routes/mod.rs,
routes/session.rs, routes/issue.rs, routes/worktree.rs, and
added crates/server/server.spec.md.
