+++ crates/server/src/lib.rs
@@ -1,58 +1,16 @@
-use axum::{
-    extract::{Path, Query, State},
-    http::StatusCode,
-    response::{
-        sse::{Event, Sse},
-        IntoResponse,
-    },
-    routing::{get, post},
-    Json, Router,
-};
-use chrono::Utc;
+use axum::{routing::{get, post}, Router};
 use db::SqliteDb;
-use futures::stream::{self, StreamExt};
-use serde::{Deserialize, Deserializer, Serialize};
-use std::{collections::{HashMap, HashSet}, convert::Infallible, path::PathBuf, sync::Arc};
+use std::{collections::{HashMap, HashSet}, path::PathBuf, sync::Arc};
 use tokio::net::TcpListener;
-use tokio_stream::wrappers::BroadcastStream;
-use types::{Issue, IssueComment, IssueStatus, Session, SessionEvent, SessionStatus};
-use uuid::Uuid;
 
-#[derive(Debug, thiserror::Error)]
-pub enum Error {
-    #[error("db error: {0}")]
-    Db(#[from] db::Error),
-    #[error("io error: {0}")]
-    Io(#[from] std::io::Error),
-    #[error("not found")]
-    NotFound,
-    #[error("bad request: {0}")]
-    BadRequest(String),
-}
+mod state;
+mod routes;
 
-pub type Result<T> = std::result::Result<T, Error>;
+pub use routes::{Error, Result};
+pub use routes::session::CreateSessionRequest;
 
-impl From<issues::Error> for Error {
-    fn from(e: issues::Error) -> Self {
-        match e {
-            issues::Error::Db(db_err) => Error::Db(db_err),
-            issues::Error::BadRequest(msg) => Error::BadRequest(msg),
-        }
-    }
-}
-
-impl IntoResponse for Error {
-    fn into_response(self) -> axum::response::Response {
-        let (status, message) = match &self {
-            Error::NotFound | Error::Db(db::Error::NotFound) => {
-                (StatusCode::NOT_FOUND, self.to_string())
-            }
-            Error::BadRequest(_) => (StatusCode::BAD_REQUEST, self.to_string()),
-            _ => (StatusCode::INTERNAL_SERVER_ERROR, self.to_string()),
-        };
-        (status, Json(serde_json::json!({ "error": message }))).into_response()
-    }
-}
+use state::AppState;
+use routes::{issue, session};
 
 pub struct ServerConfig {
     pub port: u16,
@@ -61,665 +19,24 @@ pub struct ServerConfig {
     pub model: String,
 }
 
-#[derive(Clone)]
-struct AppState {
-    db: Arc<dyn db::Db>,
-    issue_service: issues::IssueService,
-    sessions: Arc<tokio::sync::Mutex<HashMap<Uuid, tokio::sync::broadcast::Sender<SessionEvent>>>>,
-    msg_senders: Arc<tokio::sync::Mutex<HashMap<Uuid, tokio::sync::mpsc::Sender<String>>>>,
-    spawning: Arc<tokio::sync::Mutex<HashSet<Uuid>>>,
-    client: Arc<dyn anthropic::AnthropicClient>,
-    tools: Vec<Arc<dyn tools::Tool>>,
-    model: String,
-}
-
-#[derive(Deserialize)]
-pub struct CreateSessionRequest {
-    pub name: Option<String>,
-    pub agent: Option<String>,
-    pub initial_message: Option<String>,
-}
-
-#[derive(Deserialize)]
-struct ListSessionsQuery {
-    status: Option<String>,
-}
-
-#[derive(Serialize)]
-struct HealthResponse {
-    status: &'static str,
-}
-
-#[derive(Deserialize)]
-struct SendMessageRequest {
-    message: String,
-}
-
-#[derive(Deserialize)]
-struct UpdateSessionStatusRequest {
-    status: String,
-}
-
-async fn health() -> Json<HealthResponse> {
-    Json(HealthResponse { status: "ok" })
-}
-
-fn spawn_harness_sync(
-    state: &AppState,
-    session: Session,
-    issue_id: Option<String>,
-) -> tokio::sync::mpsc::Sender<String> {
-    let (tx, _rx) = tokio::sync::broadcast::channel::<SessionEvent>(256);
-    let (msg_tx, msg_rx) = tokio::sync::mpsc::channel::<String>(16);
-
-    let sessions_map = Arc::clone(&state.sessions);
-    let msg_senders_map = Arc::clone(&state.msg_senders);
-    let spawning_set = Arc::clone(&state.spawning);
-    let db = Arc::clone(&state.db);
-    let client = Arc::clone(&state.client);
-    let session_clone = session.clone();
-    let event_tx = tx.clone();
-    let msg_tx_ret = msg_tx.clone();
-    let tools = state.tools.clone();
-    let model = state.model.clone();
-
-    let config = harness::HarnessConfig {
-        session: session_clone.clone(),
-        model,
-        tools,
-        git_root: None,
-    };
-
-    let session_id = session_clone.id;
-
-    let issue_watcher = issue_id.map(|id| {
-        let mut rx = tx.subscribe();
-        let db_watch = Arc::clone(&db);
-        tokio::spawn(async move {
-            let mut current_turn_text = String::new();
-            let mut last_turn_text = String::new();
-
-            while let Ok(event) = rx.recv().await {
-                match event {
-                    SessionEvent::ContentBlockDelta {
-                        delta: types::ContentBlockDelta::TextDelta { text },
-                        ..
-                    } => {
-                        current_turn_text.push_str(&text);
-                    }
-                    SessionEvent::TurnDone { .. } if !current_turn_text.is_empty() => {
-                        last_turn_text = std::mem::take(&mut current_turn_text);
-                    }
-                    SessionEvent::SessionDone { .. } => {
-                        if let Ok(mut issue) = db_watch.get_issue(id.clone()).await {
-                            if !last_turn_text.is_empty() {
-                                let author = issue.assignee.clone()
-                                    .unwrap_or_else(|| "agent".to_string());
-                                issue.comments.push(IssueComment {
-                                    author,
-                                    created_at: Utc::now(),
-                                    body: last_turn_text.clone(),
-                                });
-                            }
-                            issue.status = IssueStatus::Completed;
-                            issue.updated_at = Utc::now();
-                            let _ = db_watch.update_issue(&issue).await;
-                        }
-                        break;
-                    }
-                    SessionEvent::Error { message } => {
-                        if let Ok(mut issue) = db_watch.get_issue(id.clone()).await {
-                            issue.comments.push(IssueComment {
-                                author: "system".to_string(),
-                                created_at: Utc::now(),
-                                body: message.clone(),
-                            });
-                            issue.status = IssueStatus::Failed;
-                            issue.updated_at = Utc::now();
-                            let _ = db_watch.update_issue(&issue).await;
-                        }
-                        break;
-                    }
-                    _ => {}
-                }
-            }
-        })
-    });
-
-    tokio::spawn(async move {
-        {
-            let mut smap = msg_senders_map.lock().await;
-            let mut map = sessions_map.lock().await;
-            let mut spawning = spawning_set.lock().await;
-            map.insert(session_id, event_tx.clone());
-            smap.insert(session_id, msg_tx);
-            spawning.remove(&session_id);
-        }
-
-        let db_clone = Arc::clone(&db);
-        let event_tx_clone = event_tx.clone();
-
-        if let Err(e) = harness::run(config, client, db, event_tx.clone(), msg_rx).await {
-            let _ = event_tx_clone.send(SessionEvent::Error { message: e.to_string() });
-            let _ = db_clone.update_session_status(session_id, SessionStatus::Failed).await;
-        }
-
-        {
-            let mut map = sessions_map.lock().await;
-            map.remove(&session_id);
-        }
-        {
-            let mut smap = msg_senders_map.lock().await;
-            smap.remove(&session_id);
-        }
-
-        drop(issue_watcher);
-    });
-
-    msg_tx_ret
-}
-
-async fn create_session(
-    State(state): State<AppState>,
-    Json(req): Json<CreateSessionRequest>,
-) -> std::result::Result<(StatusCode, Json<Session>), Error> {
-    let now = Utc::now();
-    let has_message = req
-        .initial_message
-        .as_deref()
-        .map(|m| !m.is_empty())
-        .unwrap_or(false);
-
-    let initial_message = req.initial_message.unwrap_or_default();
-
-    let session = Session {
-        id: Uuid::new_v4(),
-        name: req.name.unwrap_or_else(|| "unnamed".to_string()),
-        status: SessionStatus::Created,
-        agent: req.agent,
-        created_at: now,
-        updated_at: now,
-    };
-    state.db.create_session(&session).await?;
-
-    if has_message {
-        let msg_tx = spawn_harness_sync(&state, session.clone(), None);
-        msg_tx.send(initial_message).await.ok();
-    }
-
-    Ok((StatusCode::CREATED, Json(session)))
-}
-
-
-async fn list_sessions(
-    State(state): State<AppState>,
-    Query(params): Query<ListSessionsQuery>,
-) -> std::result::Result<Json<Vec<Session>>, Error> {
-    let status = params
-        .status
-        .as_deref()
-        .map(|s| s.parse::<SessionStatus>().map_err(Error::BadRequest))
-        .transpose()?;
-    let sessions = state.db.list_sessions(status).await?;
-    Ok(Json(sessions))
-}
-
-async fn get_session(
-    State(state): State<AppState>,
-    Path(id): Path<Uuid>,
-) -> std::result::Result<Json<Session>, Error> {
-    let session = state.db.get_session(id).await?;
-    Ok(Json(session))
-}
-
-fn event_from(ev: &SessionEvent) -> Event {
-    Event::default().data(serde_json::to_string(ev).unwrap_or_default())
-}
-
-#[derive(Deserialize)]
-struct SessionEventsQuery {
-    last_turns: Option<usize>,
-}
-
-async fn session_events(
-    State(state): State<AppState>,
-    Path(id): Path<Uuid>,
-    Query(params): Query<SessionEventsQuery>,
-) -> Sse<impl futures::Stream<Item = std::result::Result<Event, Infallible>>> {
-    // Subscribe BEFORE reading history to avoid race
-    let live_rx = {
-        let map = state.sessions.lock().await;
-        map.get(&id).map(|tx| tx.subscribe())
-    };
-
-    // Fetch session to check status
-    let session = state.db.get_session(id).await;
-
-    // Build historical events
-    let mut history: Vec<SessionEvent> = Vec::new();
-
-    if let Ok(ref sess) = session {
-        if let Ok(turns) = state.db.list_turns(id).await {
-            // Apply last_turns filter: if last_turns=0, skip all; if last_turns=N, take last N; if absent, take all
-            let turns_to_replay: Vec<_> = match params.last_turns {
-                Some(0) => vec![],
-                Some(n) => turns.iter().rev().take(n).rev().cloned().collect(),
-                None => turns.clone(),
-            };
-
-            for turn in &turns_to_replay {
-                history.push(SessionEvent::TurnStarted { turn: turn.clone() });
-                if let Ok(blocks) = state.db.list_content_blocks(turn.id).await {
-                    for (i, (_role, block)) in blocks.into_iter().enumerate() {
-                        let index = i as u32;
-                        if let types::ContentBlock::Text { ref text } = block {
-                            history.push(SessionEvent::ContentBlockDelta {
-                                turn_id: turn.id,
-                                index,
-                                delta: types::ContentBlockDelta::TextDelta { text: text.clone() },
-                            });
-                        }
-                        history.push(SessionEvent::ContentBlockDone {
-                            turn_id: turn.id,
-                            index,
-                            block,
-                        });
-                    }
-                }
-                history.push(SessionEvent::TurnDone { turn_id: turn.id });
-            }
-        }
-
-        match sess.status {
-            SessionStatus::Completed | SessionStatus::Failed | SessionStatus::Cancelled => {
-                history.push(SessionEvent::SessionDone { session_id: id });
-            }
-            _ => {}
-        }
-    }
-
-    let history_stream = stream::iter(history)
-        .map(|event| Ok::<Event, Infallible>(event_from(&event)));
-
-    let live_stream = if let Some(rx) = live_rx {
-        futures::future::Either::Left(
-            BroadcastStream::new(rx)
-                .filter_map(|result| async move {
-                    match result {
-                        Ok(event) => Some(event),
-                        Err(_lagged) => Some(SessionEvent::Error {
-                            message: "stream lagged".into(),
-                        }),
-                    }
-                })
-                .map(|event| Ok::<Event, Infallible>(event_from(&event))),
-        )
-    } else {
-        futures::future::Either::Right(stream::empty::<std::result::Result<Event, Infallible>>())
-    };
-
-    Sse::new(history_stream.chain(live_stream))
-}
-
-/// Send a message to a session. Works on `created`, `running`, and `completed` sessions.
-///
-/// - If the session has an active harness (sender in map): queue the message directly.
-/// - If the session exists in DB but has no active harness (`created` status): spawn a
-///   new harness and send the message to it.
-/// - If the session does not exist: 404.
-async fn send_message(
-    State(state): State<AppState>,
-    Path(id): Path<Uuid>,
-    Json(req): Json<SendMessageRequest>,
-) -> std::result::Result<StatusCode, Error> {
-    // Fast path: sender already exists (running session with live harness).
-    {
-        let senders = state.msg_senders.lock().await;
-        if let Some(tx) = senders.get(&id) {
-            tx.send(req.message).await.ok();
-            return Ok(StatusCode::OK);
-        }
-    }
-
-    // Verify the session exists in DB.
-    let session = state.db.get_session(id).await.map_err(|e| match e {
-        db::Error::NotFound => Error::NotFound,
-        other => Error::Db(other),
-    })?;
-
-    // For `created`, `running`, and `completed` sessions, spawn a harness (if not already
-    // running) and deliver the message. `completed` sessions resume from full DB history
-    // so no in-memory state is required.
-    //
-    // Guard against two concurrent requests both reaching the spawn path simultaneously
-    // using the `spawning` set as an atomic "already in flight" guard.
-    match session.status {
-        SessionStatus::Created | SessionStatus::Running | SessionStatus::Completed => {
-            // Acquire both locks together to make the check-then-act atomic.
-            let mut spawning = state.spawning.lock().await;
-            let senders = state.msg_senders.lock().await;
-
-            if let Some(tx) = senders.get(&id) {
-                // A concurrent request beat us; the harness is now live.
-                let tx = tx.clone();
-                drop(senders);
-                drop(spawning);
-                tx.send(req.message).await.ok();
-                return Ok(StatusCode::OK);
-            }
-
-            if spawning.contains(&id) {
-                // Another request is already spawning. Return OK; the message will be
-                // processed once the harness registers its sender and receives messages.
-                drop(senders);
-                drop(spawning);
-                // Spin-wait briefly for the sender to appear.
-                for _ in 0..40 {
-                    tokio::time::sleep(tokio::time::Duration::from_millis(5)).await;
-                    let senders = state.msg_senders.lock().await;
-                    if let Some(tx) = senders.get(&id) {
-                        tx.send(req.message).await.ok();
-                        return Ok(StatusCode::OK);
-                    }
-                }
-                // Still not available — best-effort: return OK (message may be lost but
-                // the client gets a success status rather than a confusing error).
-                return Ok(StatusCode::OK);
-            }
-
-            // We are the first to reach this point. Mark as spawning and do it.
-            spawning.insert(id);
-            drop(senders);
-            drop(spawning);
-
-            let msg_tx = spawn_harness_sync(&state, session, None);
-            msg_tx.send(req.message).await.ok();
-            Ok(StatusCode::OK)
-        }
-        // `failed` and `cancelled` are terminal states that cannot accept new messages.
-        // The caller must explicitly reopen the session (e.g. via `issue reopen`) first.
-        SessionStatus::Failed => Err(Error::BadRequest(
-            "session is in failed state and cannot accept messages; reopen it first".into(),
-        )),
-        SessionStatus::Cancelled => Err(Error::BadRequest(
-            "session is in cancelled state and cannot accept messages; reopen it first".into(),
-        )),
-    }
-}
-
-async fn update_session_status(
-    State(state): State<AppState>,
-    Path(id): Path<Uuid>,
-    Json(req): Json<UpdateSessionStatusRequest>,
-) -> std::result::Result<Json<Session>, Error> {
-    let new_status = req
-        .status
-        .parse::<SessionStatus>()
-        .map_err(Error::BadRequest)?;
-    state.db.update_session_status(id, new_status).await?;
-    let session = state.db.get_session(id).await?;
-    Ok(Json(session))
-}
-
-
-fn generate_issue_id() -> String {
-    const ALPHABET: &[u8] = b"abcdefghijklmnopqrstuvwxyz0123456789";
-    let id = Uuid::new_v4();
-    let bytes = id.as_bytes();
-    (0..4).map(|i| ALPHABET[(bytes[i] as usize) % ALPHABET.len()] as char).collect()
-}
-
-/// Convert a title string into a URL-safe slug.
-/// - Lowercase all characters
-/// - Replace any run of non-alphanumeric characters with a single `-`
-/// - Trim leading/trailing `-`
-fn slugify(title: &str) -> String {
-    let lower = title.to_lowercase();
-    // Replace runs of non-alphanumeric characters with a single dash
-    let mut result = String::new();
-    let mut in_sep = false;
-    for ch in lower.chars() {
-        if ch.is_alphanumeric() {
-            result.push(ch);
-            in_sep = false;
-        } else if !in_sep {
-            result.push('-');
-            in_sep = true;
-        }
-    }
-    // Trim leading and trailing dashes
-    result.trim_matches('-').to_string()
-}
-
-#[derive(Deserialize)]
-struct CreateIssueRequest {
-    title: String,
-    body: String,
-    assignee: Option<String>,
-    parent_id: Option<String>,
-    blocked_on: Option<Vec<String>>,
-    branch: Option<String>,
-}
-
-// Wraps a present JSON field (including null) in Some, leaving absent fields as None.
-// Used with #[serde(default, deserialize_with = "deserialize_some")] to distinguish
-// "field absent" (None) from "field explicitly null" (Some(None)).
-fn deserialize_some<'de, T, D>(deserializer: D) -> std::result::Result<Option<T>, D::Error>
-where
-    T: Deserialize<'de>,
-    D: Deserializer<'de>,
-{
-    Deserialize::deserialize(deserializer).map(Some)
-}
-
-#[derive(Deserialize)]
-struct EditIssueRequest {
-    title: Option<String>,
-    body: Option<String>,
-    #[serde(default, deserialize_with = "deserialize_some")]
-    assignee: Option<Option<String>>,
-    #[serde(default, deserialize_with = "deserialize_some")]
-    parent_id: Option<Option<String>>,
-    blocked_on: Option<Vec<String>>,
-    branch: Option<String>,
-}
-
-#[derive(Deserialize)]
-struct ListIssuesQuery {
-    status: Option<String>,
-    assignee: Option<String>,
-    parent_id: Option<String>,
-    blocked_on: Option<String>,
-}
-
-#[derive(Deserialize)]
-struct AddCommentRequest {
-    author: String,
-    body: String,
-}
-
-#[derive(Deserialize)]
-struct CompleteIssueRequest {
-    comment: String,
-}
-
-async fn create_issue(
-    State(state): State<AppState>,
-    Json(req): Json<CreateIssueRequest>,
-) -> std::result::Result<(StatusCode, Json<Issue>), Error> {
-    let now = Utc::now();
-    let id = generate_issue_id();
-
-    // Compute the branch:
-    // 1. Explicit branch in request → use as-is
-    // 2. parent_id provided → inherit parent's branch
-    // 3. Otherwise → generate slug from "<id>-<slugified-title>"
-    let branch = if let Some(b) = req.branch {
-        b
-    } else if let Some(ref parent_id) = req.parent_id {
-        match state.db.get_issue(parent_id.clone()).await {
-            Ok(parent) => parent.branch,
-            Err(_) => format!("{}-{}", id, slugify(&req.title)),
-        }
-    } else {
-        format!("{}-{}", id, slugify(&req.title))
-    };
-
-    let issue = Issue {
-        id,
-        title: req.title,
-        body: req.body,
-        status: IssueStatus::Open,
-        branch,
-        assignee: req.assignee,
-        session_id: None,
-        parent_id: req.parent_id,
-        blocked_on: req.blocked_on.unwrap_or_default(),
-        comments: vec![],
-        created_at: now,
-        updated_at: now,
-    };
-    state.db.create_issue(&issue).await?;
-    Ok((StatusCode::CREATED, Json(issue)))
-}
-
-async fn get_issue(
-    State(state): State<AppState>,
-    Path(id): Path<String>,
-) -> std::result::Result<Json<Issue>, Error> {
-    let issue = state.db.get_issue(id).await?;
-    Ok(Json(issue))
-}
-
-async fn list_issues(
-    State(state): State<AppState>,
-    Query(params): Query<ListIssuesQuery>,
-) -> std::result::Result<Json<Vec<Issue>>, Error> {
-    let status = params
-        .status
-        .as_deref()
-        .map(|s| s.parse::<IssueStatus>().map_err(Error::BadRequest))
-        .transpose()?;
-    let mut issues = state
-        .db
-        .list_issues(status, params.assignee, params.parent_id)
-        .await?;
-    if let Some(blocked_on_filter) = &params.blocked_on {
-        issues.retain(|i| i.blocked_on.contains(blocked_on_filter));
-    }
-    Ok(Json(issues))
-}
-
-async fn edit_issue(
-    State(state): State<AppState>,
-    Path(id): Path<String>,
-    Json(req): Json<EditIssueRequest>,
-) -> std::result::Result<Json<Issue>, Error> {
-    let mut issue = state.db.get_issue(id.clone()).await?;
-    if let Some(title) = req.title {
-        issue.title = title;
-    }
-    if let Some(body) = req.body {
-        issue.body = body;
-    }
-    if let Some(assignee_opt) = req.assignee {
-        issue.assignee = assignee_opt;
-    }
-    if let Some(parent_opt) = req.parent_id {
-        issue.parent_id = parent_opt;
-    }
-    if let Some(blocked_on) = req.blocked_on {
-        issue.blocked_on = blocked_on;
-    }
-    if let Some(branch) = req.branch {
-        issue.branch = branch;
-    }
-    issue.updated_at = Utc::now();
-    state.db.update_issue(&issue).await?;
-    Ok(Json(issue))
-}
-
-async fn add_comment(
-    State(state): State<AppState>,
-    Path(id): Path<String>,
-    Json(req): Json<AddCommentRequest>,
-) -> std::result::Result<Json<Issue>, Error> {
-    let mut issue = state.db.get_issue(id.clone()).await?;
-    issue.comments.push(IssueComment {
-        author: req.author,
-        created_at: Utc::now(),
-        body: req.body,
-    });
-    issue.updated_at = Utc::now();
-    state.db.update_issue(&issue).await?;
-    Ok(Json(issue))
-}
-
-async fn start_issue(
-    State(state): State<AppState>,
-    Path(id): Path<String>,
-) -> std::result::Result<Json<Issue>, Error> {
-    let outcome = state.issue_service.start_issue(id).await?;
-    let msg_tx = spawn_harness_sync(&state, outcome.session, Some(outcome.issue.id.clone()));
-    msg_tx.send(outcome.initial_message).await.ok();
-    Ok(Json(outcome.issue))
-}
-
-async fn complete_issue(
-    State(state): State<AppState>,
-    Path(id): Path<String>,
-    Json(req): Json<CompleteIssueRequest>,
-) -> std::result::Result<Json<Issue>, Error> {
-    let issue = state.issue_service.complete_issue(id, req.comment).await?;
-    Ok(Json(issue))
-}
-
-async fn reopen_issue(
-    State(state): State<AppState>,
-    Path(id): Path<String>,
-    body: Option<Json<ReopenIssueRequest>>,
-) -> std::result::Result<Json<Issue>, Error> {
-    let comment = body.and_then(|Json(req)| req.comment);
-    let issue = state.issue_service.reopen_issue(id, comment).await?;
-    Ok(Json(issue))
-}
-
-#[derive(Deserialize, Default)]
-struct ReopenIssueRequest {
-    comment: Option<String>,
-}
-
-/// GET /sessions/:id/last_text — return the last assistant text content block for a session.
-/// Returns JSON: {"text": "<content>"} or {"text": null} if no text content found.
-async fn session_last_text(
-    State(state): State<AppState>,
-    Path(id): Path<Uuid>,
-) -> std::result::Result<Json<serde_json::Value>, Error> {
-    // Verify the session exists (returns 404 if not)
-    let _ = state.db.get_session(id).await?;
-
-    let text = state.db.get_last_text_for_session(id).await?;
-    Ok(Json(serde_json::json!({ "text": text })))
-}
-
 fn build_router(state: AppState) -> Router {
     Router::new()
-        .route("/health", get(health))
-        .route("/sessions", post(create_session))
-        .route("/sessions", get(list_sessions))
-        .route("/sessions/:id", get(get_session))
-        .route("/sessions/:id/events", get(session_events))
-        .route("/sessions/:id/messages", post(send_message))
-        .route("/sessions/:id/status", axum::routing::patch(update_session_status))
-        .route("/sessions/:id/last_text", get(session_last_text))
-        .route("/issues", post(create_issue))
-        .route("/issues", get(list_issues))
-        .route("/issues/:id", get(get_issue))
-        .route("/issues/:id", axum::routing::patch(edit_issue))
-        .route("/issues/:id/comments", post(add_comment))
-        .route("/issues/:id/start", post(start_issue))
-        .route("/issues/:id/complete", post(complete_issue))
-        .route("/issues/:id/reopen", post(reopen_issue))
+        .route("/health", get(session::health))
+        .route("/sessions", post(session::create_session))
+        .route("/sessions", get(session::list_sessions))
+        .route("/sessions/:id", get(session::get_session))
+        .route("/sessions/:id/events", get(session::session_events))
+        .route("/sessions/:id/messages", post(session::send_message))
+        .route("/sessions/:id/status", axum::routing::patch(session::update_session_status))
+        .route("/sessions/:id/last_text", get(session::session_last_text))
+        .route("/issues", post(issue::create_issue))
+        .route("/issues", get(issue::list_issues))
+        .route("/issues/:id", get(issue::get_issue))
+        .route("/issues/:id", axum::routing::patch(issue::edit_issue))
+        .route("/issues/:id/comments", post(issue::add_comment))
+        .route("/issues/:id/start", post(issue::start_issue))
+        .route("/issues/:id/complete", post(issue::complete_issue))
+        .route("/issues/:id/reopen", post(issue::reopen_issue))
         .with_state(state)
 }
 
@@ -799,7 +116,13 @@ mod tests {
     use async_trait::async_trait;
     use axum::body::Body;
     use axum::http::{Request, StatusCode};
-    use tower::ServiceExt; // for .oneshot()
+    use tower::ServiceExt;
+    use std::sync::Arc;
+    use db::SqliteDb;
+    use uuid::Uuid;
+    use types::{Session, SessionStatus, SessionEvent, IssueStatus, IssueComment};
+    use routes::session::event_from;
+    use routes::issue::slugify;
 
     /// A minimal stub AnthropicClient defined locally in server tests.
     /// Does NOT use harness::StubClient — server tests must not depend on harness internals.
@@ -969,7 +292,6 @@ mod tests {
     async fn test_list_sessions_after_create() {
         let app = test_app().await;
 
-        // Create a session
         let create_resp = app
             .clone()
             .oneshot(
@@ -979,7 +301,6 @@ mod tests {
             .unwrap();
         assert_eq!(create_resp.status(), StatusCode::CREATED);
 
-        // List sessions
         let list_resp = app
             .oneshot(
                 Request::builder()
@@ -1001,7 +322,6 @@ mod tests {
     async fn test_get_session_by_id() {
         let app = test_app().await;
 
-        // Create a session and capture its id
         let create_resp = app
             .clone()
             .oneshot(
@@ -1013,7 +333,6 @@ mod tests {
         let created: serde_json::Value = serde_json::from_slice(&body).unwrap();
         let id = created["id"].as_str().unwrap().to_owned();
 
-        // Fetch by id
         let get_resp = app
             .oneshot(
                 Request::builder()
@@ -1052,7 +371,6 @@ mod tests {
     async fn test_list_sessions_filter_by_status() {
         let app = test_app().await;
 
-        // Create a session (no initial_message -> Created status)
         app.clone()
             .oneshot(
                 create_session_req(serde_json::json!({"name": "created-sess"})).await,
@@ -1060,7 +378,6 @@ mod tests {
             .await
             .unwrap();
 
-        // Filter by "created" — should find it
         let resp = app
             .clone()
             .oneshot(
@@ -1082,13 +399,11 @@ mod tests {
     async fn test_list_sessions_filter_by_status_no_match() {
         let app = test_app().await;
 
-        // Create a session with Created status
         app.clone()
             .oneshot(create_session_req(serde_json::json!({})).await)
             .await
             .unwrap();
 
-        // Filter by "completed" — should be empty
         let resp = app
             .oneshot(
                 Request::builder()
@@ -1173,7 +488,6 @@ mod tests {
 
     // --- POST /sessions/:id/messages ---
 
-    /// Sending to a nonexistent session returns 404.
     #[tokio::test]
     async fn test_send_message_to_nonexistent_session_returns_404() {
         let app = test_app().await;
@@ -1195,12 +509,10 @@ mod tests {
         assert_eq!(resp.status(), StatusCode::NOT_FOUND);
     }
 
-    /// Sending to a `created` session (no initial message) works: spawns harness and returns 200.
     #[tokio::test]
     async fn test_send_message_to_created_session_returns_200() {
         let app = test_app().await;
 
-        // Create a session without an initial message → status = created, no harness
         let create_resp = app
             .clone()
             .oneshot(create_session_req(serde_json::json!({"name": "idle"})).await)
@@ -1210,7 +522,6 @@ mod tests {
         let created: serde_json::Value = serde_json::from_slice(&body).unwrap();
         let id = created["id"].as_str().unwrap().to_owned();
 
-        // Send a message to the created session
         let resp = app
             .oneshot(
                 Request::builder()
@@ -1227,12 +538,10 @@ mod tests {
         assert_eq!(resp.status(), StatusCode::OK);
     }
 
-    /// Sending to a running session (with initial message that spawned a harness) returns 200.
     #[tokio::test]
     async fn test_send_message_to_running_session_returns_200() {
         let (app, _state) = test_app_with_state().await;
 
-        // Create a session with an initial message → harness spawned
         let create_resp = app
             .clone()
             .oneshot(
@@ -1248,7 +557,6 @@ mod tests {
         let created: serde_json::Value = serde_json::from_slice(&body).unwrap();
         let id = created["id"].as_str().unwrap().to_owned();
 
-        // Give the spawned harness task a moment to register itself in the map
         tokio::time::sleep(tokio::time::Duration::from_millis(50)).await;
 
         let resp = app
@@ -1268,13 +576,10 @@ mod tests {
         assert_eq!(resp.status(), StatusCode::OK);
     }
 
-    /// Two concurrent send_message calls on a Created session must not spawn two harnesses.
-    /// After both calls, exactly one broadcast sender should exist in the sessions map.
     #[tokio::test]
     async fn test_concurrent_send_message_spawns_only_one_harness() {
         let (app, state) = test_app_with_state().await;
 
-        // Create a session WITHOUT an initial message → Created status, no harness
         let create_resp = app
             .clone()
             .oneshot(create_session_req(serde_json::json!({"name": "race-test"})).await)
@@ -1285,7 +590,6 @@ mod tests {
         let id = created["id"].as_str().unwrap().to_owned();
         let session_id: Uuid = id.parse().unwrap();
 
-        // Fire two concurrent send_message requests
         let app1 = app.clone();
         let app2 = app.clone();
         let id1 = id.clone();
@@ -1316,10 +620,8 @@ mod tests {
         assert_eq!(r1.unwrap().status(), StatusCode::OK);
         assert_eq!(r2.unwrap().status(), StatusCode::OK);
 
-        // Wait a moment for the spawned task to register
         tokio::time::sleep(tokio::time::Duration::from_millis(50)).await;
 
-        // Exactly one sender must exist in the sessions (broadcast) map
         let sessions = state.sessions.lock().await;
         let sender_count = if sessions.contains_key(&session_id) { 1 } else { 0 };
         assert_eq!(
@@ -1332,11 +634,14 @@ mod tests {
 
     #[tokio::test]
     async fn test_event_from_serializes_session_event() {
+        use axum::response::{sse::Sse, IntoResponse};
+        use std::convert::Infallible;
+
         let session_id = Uuid::new_v4();
         let ev = SessionEvent::SessionDone { session_id };
         let sse_event = event_from(&ev);
 
-        let stream = futures::stream::once(async move { Ok::<Event, Infallible>(sse_event) });
+        let stream = futures::stream::once(async move { Ok::<_, Infallible>(sse_event) });
         let resp = Sse::new(stream).into_response();
         let body = response_body_bytes(resp).await;
         let raw = std::str::from_utf8(&body).unwrap();
@@ -1445,12 +750,10 @@ mod tests {
 
     // --- GET /sessions/:id/events?last_turns=N ---
 
-    /// Test that last_turns=0 skips all history and only emits SessionDone for terminal sessions.
     #[tokio::test]
     async fn test_session_events_last_turns_zero_skips_history() {
         let (app, state) = test_app_with_state().await;
 
-        // Create a session and manually mark it completed with some turns
         let session = Session {
             id: Uuid::new_v4(),
             name: "turns-test".into(),
@@ -1460,8 +763,7 @@ mod tests {
             updated_at: chrono::Utc::now(),
         };
         state.db.create_session(&session).await.unwrap();
-        
-        // Create 3 turns
+
         for _ in 0..3 {
             let turn = types::Turn {
                 id: Uuid::new_v4(),
@@ -1481,14 +783,13 @@ mod tests {
                 .await
                 .unwrap();
         }
-        
+
         state
             .db
             .update_session_status(session.id, SessionStatus::Completed)
             .await
             .unwrap();
 
-        // Request with last_turns=0
         let resp = app
             .oneshot(
                 Request::builder()
@@ -1503,7 +804,6 @@ mod tests {
         let body = response_body_bytes(resp).await;
         let raw = std::str::from_utf8(&body).unwrap();
 
-        // Should NOT have any TurnStarted events
         let has_turn_started = raw.lines().filter(|l| l.starts_with("data: ")).any(|l| {
             let json = &l["data: ".len()..];
             serde_json::from_str::<SessionEvent>(json)
@@ -1511,8 +811,7 @@ mod tests {
                 .unwrap_or(false)
         });
         assert!(!has_turn_started, "last_turns=0 should skip all history turns");
-        
-        // Should still have SessionDone
+
         let has_session_done = raw.lines().filter(|l| l.starts_with("data: ")).any(|l| {
             let json = &l["data: ".len()..];
             serde_json::from_str::<SessionEvent>(json)
@@ -1522,7 +821,6 @@ mod tests {
         assert!(has_session_done, "completed session should still emit SessionDone");
     }
 
-    /// Test that last_turns=1 emits only the last turn's events.
     #[tokio::test]
     async fn test_session_events_last_turns_one_emits_last_turn() {
         let (app, state) = test_app_with_state().await;
@@ -1536,8 +834,7 @@ mod tests {
             updated_at: chrono::Utc::now(),
         };
         state.db.create_session(&session).await.unwrap();
-        
-        // Create 3 turns with distinct content
+
         let mut turn_ids = Vec::new();
         for i in 0..3 {
             let turn = types::Turn {
@@ -1559,14 +856,13 @@ mod tests {
                 .await
                 .unwrap();
         }
-        
+
         state
             .db
             .update_session_status(session.id, SessionStatus::Completed)
             .await
             .unwrap();
 
-        // Request with last_turns=1
         let resp = app
             .oneshot(
                 Request::builder()
@@ -1581,7 +877,6 @@ mod tests {
         let body = response_body_bytes(resp).await;
         let raw = std::str::from_utf8(&body).unwrap();
 
-        // Count TurnStarted events
         let turn_started_count = raw
             .lines()
             .filter(|l| l.starts_with("data: "))
@@ -1594,14 +889,11 @@ mod tests {
             .count();
         assert_eq!(turn_started_count, 1, "last_turns=1 should emit exactly 1 turn");
 
-        // Should contain the last turn's content
         assert!(raw.contains("turn-2"), "should contain last turn's content");
-        // Should NOT contain earlier turns' content
         assert!(!raw.contains("turn-0"), "should not contain first turn's content");
         assert!(!raw.contains("turn-1"), "should not contain second turn's content");
     }
 
-    /// Test that absent last_turns param emits all history (current behavior).
     #[tokio::test]
     async fn test_session_events_no_last_turns_emits_all_history() {
         let (app, state) = test_app_with_state().await;
@@ -1615,8 +907,7 @@ mod tests {
             updated_at: chrono::Utc::now(),
         };
         state.db.create_session(&session).await.unwrap();
-        
-        // Create 3 turns
+
         for i in 0..3 {
             let turn = types::Turn {
                 id: Uuid::new_v4(),
@@ -1636,14 +927,13 @@ mod tests {
                 .await
                 .unwrap();
         }
-        
+
         state
             .db
             .update_session_status(session.id, SessionStatus::Completed)
             .await
             .unwrap();
 
-        // Request without last_turns param
         let resp = app
             .oneshot(
                 Request::builder()
@@ -1658,19 +948,15 @@ mod tests {
         let body = response_body_bytes(resp).await;
         let raw = std::str::from_utf8(&body).unwrap();
 
-        // Should contain all turns' content
         assert!(raw.contains("turn-0"), "should contain first turn's content");
         assert!(raw.contains("turn-1"), "should contain second turn's content");
         assert!(raw.contains("turn-2"), "should contain last turn's content");
     }
 
-    /// A Completed session (with active harness still waiting for messages) accepts follow-up
-    /// messages and returns 200. This is the in-process multi-turn case.
     #[tokio::test]
     async fn test_completed_session_sender_still_alive() {
         let (app, state) = test_app_with_state().await;
 
-        // Create a session with an initial message so a harness is spawned
         let create_resp = app
             .clone()
             .oneshot(
@@ -1687,7 +973,6 @@ mod tests {
         let id = created["id"].as_str().unwrap().to_owned();
         let session_id: Uuid = id.parse().unwrap();
 
-        // Wait for the session to reach Completed status in the DB
         let deadline = tokio::time::Instant::now() + tokio::time::Duration::from_secs(3);
         loop {
             tokio::time::sleep(tokio::time::Duration::from_millis(20)).await;
@@ -1700,9 +985,6 @@ mod tests {
             }
         }
 
-        // The harness is still alive (waiting for next message in its loop), so the
-        // msg sender is still in the map — this is the normal in-process case.
-        // POST /sessions/:id/messages must return 200 via the fast path (sender in map).
         let resp = app
             .oneshot(
                 Request::builder()
@@ -1724,13 +1006,10 @@ mod tests {
         );
     }
 
-    /// POST /sessions/:id/messages on a `Completed` session with no active sender
-    /// must return 200 OK and spawn a fresh harness.
     #[tokio::test]
     async fn test_send_message_to_completed_session_returns_200() {
         let (app, state) = test_app_with_state().await;
 
-        // Directly insert a completed session into the DB (no harness ever spawned)
         let session = Session {
             id: Uuid::new_v4(),
             name: "completed-no-harness".into(),
@@ -1746,7 +1025,6 @@ mod tests {
             .await
             .unwrap();
 
-        // Confirm no sender in the map
         {
             let senders = state.msg_senders.lock().await;
             assert!(!senders.contains_key(&session.id));
@@ -1772,8 +1050,6 @@ mod tests {
         );
     }
 
-    /// POST /sessions/:id/messages on a `Failed` session must return a non-2xx status
-    /// with an error body (sessions in Failed state cannot accept messages).
     #[tokio::test]
     async fn test_send_message_to_failed_session_returns_bad_request() {
         let (app, state) = test_app_with_state().await;
@@ -1822,8 +1098,6 @@ mod tests {
         );
     }
 
-    /// POST /sessions/:id/messages on a `Cancelled` session must return a non-2xx status
-    /// with an error body (sessions in Cancelled state cannot accept messages).
     #[tokio::test]
     async fn test_send_message_to_cancelled_session_returns_bad_request() {
         let (app, state) = test_app_with_state().await;
@@ -2117,7 +1391,6 @@ mod tests {
         let created: serde_json::Value = serde_json::from_slice(&body).unwrap();
         let id = created["id"].as_str().unwrap();
 
-        // Patch only title — parent_id absent from request should leave it unchanged
         let edit_resp = app
             .oneshot(issue_req("PATCH", &format!("/issues/{id}"), serde_json::json!({
                 "title": "Renamed"
@@ -2190,7 +1463,6 @@ mod tests {
     #[tokio::test]
     async fn test_start_issue_on_non_open_issue_returns_400() {
         let app = test_app().await;
-        // Create with assignee so the first start succeeds
         let create_resp = app
             .clone()
             .oneshot(issue_req("POST", "/issues", serde_json::json!({
@@ -2202,7 +1474,6 @@ mod tests {
         let created: serde_json::Value = serde_json::from_slice(&body).unwrap();
         let id = created["id"].as_str().unwrap();
 
-        // First start (should succeed and move to running)
         app.clone()
             .oneshot(
                 Request::builder()
@@ -2214,7 +1485,6 @@ mod tests {
             .await
             .unwrap();
 
-        // Second start should fail (already running)
         let second_start = app
             .oneshot(
                 Request::builder()
@@ -2293,9 +1563,6 @@ mod tests {
     async fn test_issue_auto_completes_when_session_succeeds() {
         let (app, state) = test_app_with_state().await;
 
-        // Use a non-existent agent name so the harness does not load any agent
-        // definition from disk (which would pull in include_project_config and
-        // real project hooks that depend on the working-tree state).
         let create_resp = app
             .clone()
             .oneshot(issue_req("POST", "/issues", serde_json::json!({
@@ -2318,7 +1585,6 @@ mod tests {
             .await
             .unwrap();
 
-        // Poll until the issue reaches a terminal state (harness uses TestClient → completes fast).
         let deadline = tokio::time::Instant::now() + tokio::time::Duration::from_secs(5);
         loop {
             tokio::time::sleep(tokio::time::Duration::from_millis(50)).await;
@@ -2351,7 +1617,6 @@ mod tests {
     async fn test_patch_session_status_happy_path() {
         let app = test_app().await;
 
-        // Create a session first
         let create_resp = app
             .clone()
             .oneshot(create_session_req(serde_json::json!({"name": "status-test"})).await)
@@ -2361,7 +1626,6 @@ mod tests {
         let created: serde_json::Value = serde_json::from_slice(&body).unwrap();
         let id = created["id"].as_str().unwrap().to_owned();
 
-        // Patch the status to "completed"
         let resp = patch_session_status(app, &id, serde_json::json!({"status": "completed"})).await;
         assert_eq!(resp.status(), StatusCode::OK);
 
@@ -2388,7 +1652,6 @@ mod tests {
     async fn test_patch_session_status_invalid_status_returns_400() {
         let app = test_app().await;
 
-        // Create a session first
         let create_resp = app
             .clone()
             .oneshot(create_session_req(serde_json::json!({"name": "bad-status"})).await)
@@ -2474,7 +1737,6 @@ mod tests {
 
     // ─── Orphan sweep tests ───────────────────────────────────────────────────
 
-    /// A `running` session is swept to `failed` by `orphan_sweep`.
     #[tokio::test]
     async fn test_orphan_sweep_marks_running_session_failed() {
         let state = test_state().await;
@@ -2499,8 +1761,6 @@ mod tests {
         );
     }
 
-    /// A `running` session linked to a `running` issue: both swept to `failed`,
-    /// and a system comment is appended to the issue.
     #[tokio::test]
     async fn test_orphan_sweep_marks_linked_issue_failed_with_comment() {
         let state = test_state().await;
@@ -2533,7 +1793,6 @@ mod tests {
 
         state.issue_service.orphan_sweep().await;
 
-        // Issue must be failed
         let fetched_issue = state.db.get_issue("ab12".into()).await.unwrap();
         assert_eq!(
             fetched_issue.status,
@@ -2541,7 +1800,6 @@ mod tests {
             "orphan sweep must mark the linked issue as failed"
         );
 
-        // Issue must have the system comment
         assert_eq!(
             fetched_issue.comments.len(),
             1,
@@ -2556,7 +1814,6 @@ mod tests {
         );
     }
 
-    /// `completed` and `cancelled` sessions are NOT swept.
     #[tokio::test]
     async fn test_orphan_sweep_ignores_non_running_sessions() {
         let state = test_state().await;
@@ -2584,7 +1841,6 @@ mod tests {
 
         state.issue_service.orphan_sweep().await;
 
-        // Each session must retain its original status — the sweep touches only `running`.
         for (id, original_status) in &session_ids {
             let fetched = state.db.get_session(*id).await.unwrap();
             assert_eq!(
@@ -2596,7 +1852,6 @@ mod tests {
         }
     }
 
-    /// A `running` session with NO linked issue is swept to `failed` without error.
     #[tokio::test]
     async fn test_orphan_sweep_no_linked_issue_does_not_error() {
         let state = test_state().await;
@@ -2611,7 +1866,6 @@ mod tests {
         };
         state.db.create_session(&session).await.unwrap();
 
-        // Must not panic or return an error
         state.issue_service.orphan_sweep().await;
 
         let fetched = state.db.get_session(session.id).await.unwrap();
@@ -2624,11 +1878,10 @@ mod tests {
 
     // ─── POST /issues/:id/reopen tests ───────────────────────────────────────
 
-    /// Helper: create a failed issue directly in the DB.
     async fn create_issue_with_status(state: &AppState, status: IssueStatus) -> String {
         let now = chrono::Utc::now();
         let issue = types::Issue {
-            id: generate_issue_id(),
+            id: routes::issue::generate_issue_id_for_test(),
             title: "Test issue".into(),
             body: "body".into(),
             status,
@@ -2649,8 +1902,6 @@ mod tests {
         issue.id
     }
 
-    /// POST /issues/:id/reopen on a failed issue → 200, status becomes open, session_id is null,
-    /// comments are preserved.
     #[tokio::test]
     async fn test_reopen_failed_issue_returns_200_and_transitions_to_open() {
         let (app, state) = test_app_with_state().await;
@@ -2673,14 +1924,12 @@ mod tests {
         let v: serde_json::Value = serde_json::from_slice(&body).unwrap();
         assert_eq!(v["status"], "open", "status must become open");
         assert!(v["session_id"].is_null(), "session_id must be cleared to null");
-        // Comments must be preserved
         let comments = v["comments"].as_array().unwrap();
         assert_eq!(comments.len(), 1, "existing comment must be preserved");
         assert_eq!(comments[0]["author"], "system");
         assert_eq!(comments[0]["body"], "session lost on server restart");
     }
 
-    /// POST /issues/:id/reopen on an open issue → 400 with error body.
     #[tokio::test]
     async fn test_reopen_open_issue_returns_400() {
         let (app, state) = test_app_with_state().await;
@@ -2712,7 +1961,6 @@ mod tests {
         );
     }
 
-    /// POST /issues/:id/reopen on a running issue → 400 with error body.
     #[tokio::test]
     async fn test_reopen_running_issue_returns_400() {
         let (app, state) = test_app_with_state().await;
@@ -2740,17 +1988,11 @@ mod tests {
         );
     }
 
-    // --- New reopen behavior tests ---
-
-    /// POST /issues/:id/reopen on a completed issue → 200, status becomes open,
-    /// session_id is KEPT (so next start resumes the existing session history),
-    /// comments are preserved.
     #[tokio::test]
     async fn test_reopen_completed_issue_keeps_session_id() {
         let (app, state) = test_app_with_state().await;
         let id = create_issue_with_status(&state, IssueStatus::Completed).await;
 
-        // Fetch the original session_id before reopening
         let original = state.db.get_issue(id.clone()).await.unwrap();
         let original_session_id = original.session_id.expect("should have a session_id");
 
@@ -2769,24 +2011,20 @@ mod tests {
         let body = response_body_bytes(resp).await;
         let v: serde_json::Value = serde_json::from_slice(&body).unwrap();
         assert_eq!(v["status"], "open", "status must become open");
-        // session_id must be preserved for completed issues
         assert_eq!(
             v["session_id"].as_str().unwrap(),
             original_session_id.to_string(),
             "session_id must be kept when reopening a completed issue"
         );
-        // Comments preserved
         let comments = v["comments"].as_array().unwrap();
         assert_eq!(comments.len(), 1, "existing comment must be preserved");
     }
 
-    /// POST /issues/:id/reopen on a failed issue → session_id is cleared.
     #[tokio::test]
     async fn test_reopen_failed_issue_clears_session_id() {
         let (app, state) = test_app_with_state().await;
         let id = create_issue_with_status(&state, IssueStatus::Failed).await;
 
-        // Verify the issue has a session_id before reopening
         let original = state.db.get_issue(id.clone()).await.unwrap();
         assert!(original.session_id.is_some(), "test setup: issue should have a session_id");
 
@@ -2808,8 +2046,6 @@ mod tests {
         assert!(v["session_id"].is_null(), "session_id must be cleared for failed issues");
     }
 
-    /// POST /issues/:id/reopen with a comment in body → comment is prepended before
-    /// status transition (comment appears in the response).
     #[tokio::test]
     async fn test_reopen_with_comment_appends_comment_before_status_change() {
         let (app, state) = test_app_with_state().await;
@@ -2830,16 +2066,13 @@ mod tests {
         assert_eq!(v["status"], "open");
 
         let comments = v["comments"].as_array().unwrap();
-        // Original comment + new one
         assert_eq!(comments.len(), 2, "should have original comment plus new one");
 
-        // The new comment must be the last one and have author="user"
         let new_comment = &comments[comments.len() - 1];
         assert_eq!(new_comment["author"], "user");
         assert_eq!(new_comment["body"], "the tests were failing because of X");
     }
 
-    /// POST /issues/:id/reopen without a comment body → no extra comment added.
     #[tokio::test]
     async fn test_reopen_without_comment_adds_no_comment() {
         let (app, state) = test_app_with_state().await;
@@ -2859,19 +2092,16 @@ mod tests {
         assert_eq!(resp.status(), StatusCode::OK);
         let body = response_body_bytes(resp).await;
         let v: serde_json::Value = serde_json::from_slice(&body).unwrap();
-        // Only the original system comment from create_issue_with_status
         let comments = v["comments"].as_array().unwrap();
         assert_eq!(comments.len(), 1, "no new comment should be added when no comment provided");
     }
 
     // ─── start_issue initial message includes comment history ────────────────
 
-    /// When an issue has no comments, start_issue sends only "title\n\nbody" to the agent.
     #[tokio::test]
     async fn test_start_issue_initial_message_no_comments() {
         use std::sync::Mutex;
 
-        // A client that records the user message text it receives
         struct CapturingClient {
             captured: Arc<Mutex<Vec<String>>>,
         }
@@ -2882,7 +2112,6 @@ mod tests {
                 &self,
                 request: anthropic::MessageRequest,
             ) -> anthropic::Result<anthropic::MessageResponse> {
-                // The first user message is the initial_message we care about
                 if let Some((_, blocks)) = request.messages.iter().find(|(role, _)| {
                     matches!(role, types::Role::User)
                 }) {
@@ -2916,7 +2145,6 @@ mod tests {
         };
         let app = build_router(state.clone());
 
-        // Create issue with NO comments
         let create_resp = app
             .clone()
             .oneshot(issue_req("POST", "/issues", serde_json::json!({
@@ -2930,7 +2158,6 @@ mod tests {
         let created: serde_json::Value = serde_json::from_slice(&body).unwrap();
         let issue_id = created["id"].as_str().unwrap().to_string();
 
-        // Start the issue
         app.clone()
             .oneshot(
                 Request::builder()
@@ -2942,7 +2169,6 @@ mod tests {
             .await
             .unwrap();
 
-        // Wait for the harness to call complete()
         let deadline = tokio::time::Instant::now() + tokio::time::Duration::from_secs(5);
         loop {
             tokio::time::sleep(tokio::time::Duration::from_millis(20)).await;
@@ -2966,7 +2192,6 @@ mod tests {
         );
     }
 
-    /// When an issue has comments, start_issue appends the comment history to the initial message.
     #[tokio::test]
     async fn test_start_issue_initial_message_includes_comments() {
         use std::sync::Mutex;
@@ -3014,7 +2239,6 @@ mod tests {
         };
         let app = build_router(state.clone());
 
-        // Create issue
         let create_resp = app
             .clone()
             .oneshot(issue_req("POST", "/issues", serde_json::json!({
@@ -3028,7 +2252,6 @@ mod tests {
         let created: serde_json::Value = serde_json::from_slice(&body).unwrap();
         let issue_id = created["id"].as_str().unwrap().to_string();
 
-        // Add two comments via the API
         app.clone()
             .oneshot(issue_req("POST", &format!("/issues/{issue_id}/comments"), serde_json::json!({
                 "author": "swe",
@@ -3044,7 +2267,6 @@ mod tests {
             .await
             .unwrap();
 
-        // Start the issue — this is where the initial_message is constructed
         app.clone()
             .oneshot(
                 Request::builder()
@@ -3056,7 +2278,6 @@ mod tests {
             .await
             .unwrap();
 
-        // Wait for the harness to call complete()
         let deadline = tokio::time::Instant::now() + tokio::time::Duration::from_secs(5);
         loop {
             tokio::time::sleep(tokio::time::Duration::from_millis(20)).await;
@@ -3071,19 +2292,14 @@ mod tests {
         let msgs = captured.lock().unwrap().clone();
         let first_msg = &msgs[0];
 
-        // Must begin with title + body
         assert!(
             first_msg.starts_with("PM Task\n\nDo the thing"),
             "initial message must start with title\\n\\nbody, got: {first_msg:?}"
         );
-
-        // Must include the history section header
         assert!(
             first_msg.contains("# Issue History"),
             "initial message must contain '# Issue History', got: {first_msg:?}"
         );
-
-        // Must include both comments' bodies
         assert!(
             first_msg.contains("Slice 1 done"),
             "initial message must contain first comment body, got: {first_msg:?}"
@@ -3092,8 +2308,6 @@ mod tests {
             first_msg.contains("session lost on server restart"),
             "initial message must contain second comment body, got: {first_msg:?}"
         );
-
-        // Must include authors
         assert!(
             first_msg.contains("swe"),
             "initial message must include comment author 'swe', got: {first_msg:?}"
@@ -3104,7 +2318,6 @@ mod tests {
         );
     }
 
-    /// POST /issues/zzzz/reopen → 404.
     #[tokio::test]
     async fn test_reopen_nonexistent_issue_returns_404() {
         let app = test_app().await;
@@ -3125,13 +2338,10 @@ mod tests {
 
     // ─── issue_watcher: final turn text posted as comment ─────────────────────
 
-    /// Test 1: start_issue → session runs → SessionDone; issue must have at least one comment
-    /// with author == assignee containing the stub response text ("stub response").
     #[tokio::test]
     async fn test_issue_watcher_posts_final_turn_as_comment_on_session_done() {
         let (app, state) = test_app_with_state().await;
 
-        // Create an issue with an assignee
         let create_resp = app
             .clone()
             .oneshot(issue_req("POST", "/issues", serde_json::json!({
@@ -3145,7 +2355,6 @@ mod tests {
         let created: serde_json::Value = serde_json::from_slice(&body).unwrap();
         let issue_id = created["id"].as_str().unwrap().to_string();
 
-        // Start the issue — spawns harness with TestClient that returns "stub response"
         app.clone()
             .oneshot(
                 Request::builder()
@@ -3157,7 +2366,6 @@ mod tests {
             .await
             .unwrap();
 
-        // Poll until the issue reaches Completed
         let deadline = tokio::time::Instant::now() + tokio::time::Duration::from_secs(5);
         loop {
             tokio::time::sleep(tokio::time::Duration::from_millis(50)).await;
@@ -3173,7 +2381,6 @@ mod tests {
         let issue = state.db.get_issue(issue_id.clone()).await.unwrap();
         assert_eq!(issue.status, IssueStatus::Completed);
 
-        // There must be at least one comment with author == assignee
         let agent_comments: Vec<_> = issue.comments.iter()
             .filter(|c| c.author == "swe-agent")
             .collect();
@@ -3183,7 +2390,6 @@ mod tests {
             issue.comments
         );
 
-        // The comment body must contain the stub response text
         let has_stub_text = agent_comments.iter().any(|c| c.body.contains("stub response"));
         assert!(
             has_stub_text,
@@ -3192,13 +2398,10 @@ mod tests {
         );
     }
 
-    /// Test 2: on Error event, issue must have a comment with author == "system" containing
-    /// the error message, and status must be Failed.
     #[tokio::test]
     async fn test_issue_watcher_posts_error_as_system_comment_on_error() {
         use async_trait::async_trait;
 
-        // A client that always returns an error
         struct ErrorClient;
 
         #[async_trait]
@@ -3229,7 +2432,6 @@ mod tests {
         };
         let app = build_router(state.clone());
 
-        // Create an issue with an assignee
         let create_resp = app
             .clone()
             .oneshot(issue_req("POST", "/issues", serde_json::json!({
@@ -3243,7 +2445,6 @@ mod tests {
         let created: serde_json::Value = serde_json::from_slice(&body).unwrap();
         let issue_id = created["id"].as_str().unwrap().to_string();
 
-        // Start the issue — the ErrorClient will cause a harness error
         app.clone()
             .oneshot(
                 Request::builder()
@@ -3255,7 +2456,6 @@ mod tests {
             .await
             .unwrap();
 
-        // Poll until the issue reaches Failed
         let deadline = tokio::time::Instant::now() + tokio::time::Duration::from_secs(5);
         loop {
             tokio::time::sleep(tokio::time::Duration::from_millis(50)).await;
@@ -3271,7 +2471,6 @@ mod tests {
         let issue = state.db.get_issue(issue_id.clone()).await.unwrap();
         assert_eq!(issue.status, IssueStatus::Failed);
 
-        // Must have a system comment containing the error message
         let system_comments: Vec<_> = issue.comments.iter()
             .filter(|c| c.author == "system")
             .collect();
@@ -3281,7 +2480,6 @@ mod tests {
             issue.comments
         );
 
-        // Comment body must contain the error text
         let has_error_text = system_comments.iter()
             .any(|c| c.body.contains("simulated api failure"));
         assert!(
@@ -3291,15 +2489,12 @@ mod tests {
         );
     }
 
-    /// Test 3: when a session produces multiple turns (simulate two TurnDone events before
-    /// SessionDone), only the text from the *last* turn is posted as a comment.
-    /// We verify this by manually driving events through a broadcast channel.
     #[tokio::test]
     async fn test_issue_watcher_only_posts_last_turn_text() {
-        // Build state so we can call spawn_harness_sync and inject events directly.
+        use chrono::Utc;
+
         let state = test_state().await;
 
-        // Create an issue
         let now = chrono::Utc::now();
         let issue = types::Issue {
             id: "tt01".to_string(),
@@ -3317,14 +2512,11 @@ mod tests {
         };
         state.db.create_issue(&issue).await.unwrap();
 
-        // Create a broadcast channel and subscribe the watcher manually,
-        // mirroring what spawn_harness_sync does for the issue_watcher.
         let (tx, _rx) = tokio::sync::broadcast::channel::<SessionEvent>(256);
         let mut rx = tx.subscribe();
         let db_watch = Arc::clone(&state.db);
         let issue_id = "tt01".to_string();
 
-        // Spawn the watcher logic directly (copied from the production code path)
         tokio::spawn(async move {
             let mut current_turn_text = String::new();
             let mut last_turn_text = String::new();
@@ -3367,7 +2559,6 @@ mod tests {
         let turn_id1 = Uuid::new_v4();
         let turn_id2 = Uuid::new_v4();
 
-        // Turn 1: text "first turn text"
         tx.send(SessionEvent::ContentBlockDelta {
             turn_id: turn_id1,
             index: 0,
@@ -3375,7 +2566,6 @@ mod tests {
         }).unwrap();
         tx.send(SessionEvent::TurnDone { turn_id: turn_id1 }).unwrap();
 
-        // Turn 2: text "second turn text"
         tx.send(SessionEvent::ContentBlockDelta {
             turn_id: turn_id2,
             index: 0,
@@ -3383,10 +2573,8 @@ mod tests {
         }).unwrap();
         tx.send(SessionEvent::TurnDone { turn_id: turn_id2 }).unwrap();
 
-        // Session done
         tx.send(SessionEvent::SessionDone { session_id }).unwrap();
 
-        // Wait for the watcher to process and write to DB
         let deadline = tokio::time::Instant::now() + tokio::time::Duration::from_secs(3);
         loop {
             tokio::time::sleep(tokio::time::Duration::from_millis(20)).await;
@@ -3401,7 +2589,6 @@ mod tests {
 
         let fetched = state.db.get_issue("tt01".to_string()).await.unwrap();
 
-        // Exactly one comment, containing only the last turn text
         assert_eq!(fetched.comments.len(), 1, "expected exactly 1 comment");
         let comment = &fetched.comments[0];
         assert_eq!(comment.author, "bot");
@@ -3462,7 +2649,6 @@ mod tests {
         };
         state.db.create_session(&session).await.unwrap();
 
-        // Create two turns
         let turn1 = types::Turn {
             id: Uuid::new_v4(),
             session_id: session.id,
@@ -3541,7 +2727,6 @@ mod tests {
         };
         state.db.create_turn(&turn).await.unwrap();
 
-        // First: text block
         state
             .db
             .create_content_block(
@@ -3553,7 +2738,6 @@ mod tests {
             .await
             .unwrap();
 
-        // Then: tool use block (should be skipped when looking for text)
         state
             .db
             .create_content_block(
@@ -3664,7 +2848,6 @@ mod tests {
         let v: serde_json::Value = serde_json::from_slice(&body).unwrap();
         let id = v["id"].as_str().unwrap();
         let branch = v["branch"].as_str().unwrap();
-        // branch should be "<id>-<slugified-title>"
         assert!(
             branch.starts_with(id),
             "branch '{branch}' should start with id '{id}'"
@@ -3680,7 +2863,6 @@ mod tests {
     async fn test_create_child_issue_inherits_parent_branch() {
         let app = test_app().await;
 
-        // Create parent issue (no branch → gets slug)
         let parent_resp = app
             .clone()
             .oneshot(issue_req("POST", "/issues", serde_json::json!({
@@ -3694,7 +2876,6 @@ mod tests {
         let parent_id = parent["id"].as_str().unwrap();
         let parent_branch = parent["branch"].as_str().unwrap();
 
-        // Create child issue with parent_id → should inherit parent's branch
         let child_resp = app
             .oneshot(issue_req("POST", "/issues", serde_json::json!({
                 "title": "Child Issue",
@@ -3789,10 +2970,10 @@ mod tests {
         assert_eq!(v["branch"], "feature/xyz");
     }
 
-    /// Test 4: if the session produces no text content (only tool calls), no empty comment
-    /// is posted.
     #[tokio::test]
     async fn test_issue_watcher_no_comment_when_no_text_content() {
+        use chrono::Utc;
+
         let state = test_state().await;
 
         let now = chrono::Utc::now();
@@ -3817,7 +2998,6 @@ mod tests {
         let db_watch = Arc::clone(&state.db);
         let issue_id = "nt01".to_string();
 
-        // Spawn the same watcher logic
         tokio::spawn(async move {
             let mut current_turn_text = String::new();
             let mut last_turn_text = String::new();
@@ -3859,14 +3039,9 @@ mod tests {
         let session_id = Uuid::new_v4();
         let turn_id = Uuid::new_v4();
 
-        // A turn with only a tool-use block (no ContentBlockDelta TextDelta events)
-        // TurnDone without any text accumulated
         tx.send(SessionEvent::TurnDone { turn_id }).unwrap();
-
-        // Session done with no text content
         tx.send(SessionEvent::SessionDone { session_id }).unwrap();
 
-        // Wait for watcher to write Completed
         let deadline = tokio::time::Instant::now() + tokio::time::Duration::from_secs(3);
         loop {
             tokio::time::sleep(tokio::time::Duration::from_millis(20)).await;
